+++ demo/temp/temp.js
@@ -1,69 +1,5 @@
-import Layout from '../../layouts/layout-1d.js';
-import {VirtualList} from '../../virtual-list.js';
+import '../../virtual-list-element.js';
 import {VirtualRepeater} from '../../virtual-repeater.js';
-
-class ListElement extends HTMLElement {
-  connectedCallback() {
-    this.style.display = 'block';
-    this.style.position = 'relative';
-    this.style.contain = 'strict';
-    this._updateList();
-  }
-
-  static get observedAttributes() {
-    return ['direction'];
-  }
-
-  attributeChangedCallback(name, oldVal, newVal) {
-    this[name] = newVal;
-  }
-
-  set template(template) {
-    if (!this._template) {
-      this._template = template;
-      this._updateList();
-    }
-  }
-
-  set direction(dir) {
-    this._direction = dir;
-    this._updateList();
-  }
-
-  set items(items) {
-    this._items = items;
-    this._updateList();
-  }
-
-  _updateList() {
-    if (!this._template) {
-      return;
-    }
-    if (!this._list) {
-      // Delay init to first connected as list needs to measure
-      // sizes of container and children.
-      if (!this.isConnected) {
-        return;
-      }
-      const {newChild, updateChild, recycleChild} = this._template;
-      this._layout = new Layout();
-      this._layout._overhang = 800;
-      this._layout._itemSize.y = 1000000;
-      this._list = new VirtualList({
-        container: this,
-        layout: this._layout,
-        newChild,
-        updateChild,
-        recycleChild,
-      });
-      this._layout._list = this._list;
-    }
-    this._list.items = this._items;
-    this._layout.direction = this._direction;
-  }
-}
-
-customElements.define('virtual-list', ListElement);
 
 class RepeaterElement extends HTMLElement {
   static get observedAttributes() {
